pyproject.toml
[project]
name = "langgraph-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.3.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "langchain-google-genai>=4.2.1",
]
